feat: eleven cloud models lose the name of the serializer that made them
`Page2ComponentJsonSchema` read like an implementation detail because it was one. The 2 is the version of Atlassian's
own Java class — `PageBean2X` and `PageBeanX` carry the same seven fields and no page is declared in both forms — and
`Json` and `Jackson1` name the serializer that produced the class.

Six `Page2*` lose the digit, four names lose `Json`, `JiraLabelPropertiesInputJackson1` loses the serializer, and
every retired name stays exported as a deprecated alias of its replacement, schema and type alike.

`ResolutionJson` and `TaskProgressJsonNode` keep their names: the first would collide with a `Resolution` of a
different shape, and the second is one of four models differing only in their `result` field, which wants the
treatment the pages just had rather than a new name.
