+++ src/cloud/models/jiraLabelProperties.ts
@@ -1,0 +1,60 @@
+import { z } from 'zod';
+import { apiObject, openEnum } from '#/core';
+
+export const JiraLabelPropertiesSchema = apiObject({
+  color: openEnum([
+    'GREY_LIGHTEST',
+    'GREY_LIGHTER',
+    'GREY',
+    'GREY_DARKER',
+    'GREY_DARKEST',
+    'PURPLE_LIGHTEST',
+    'PURPLE_LIGHTER',
+    'PURPLE',
+    'PURPLE_DARKER',
+    'PURPLE_DARKEST',
+    'BLUE_LIGHTEST',
+    'BLUE_LIGHTER',
+    'BLUE',
+    'BLUE_DARKER',
+    'BLUE_DARKEST',
+    'TEAL_LIGHTEST',
+    'TEAL_LIGHTER',
+    'TEAL',
+    'TEAL_DARKER',
+    'TEAL_DARKEST',
+    'GREEN_LIGHTEST',
+    'GREEN_LIGHTER',
+    'GREEN',
+    'GREEN_DARKER',
+    'GREEN_DARKEST',
+    'LIME_LIGHTEST',
+    'LIME_LIGHTER',
+    'LIME',
+    'LIME_DARKER',
+    'LIME_DARKEST',
+    'YELLOW_LIGHTEST',
+    'YELLOW_LIGHTER',
+    'YELLOW',
+    'YELLOW_DARKER',
+    'YELLOW_DARKEST',
+    'ORANGE_LIGHTEST',
+    'ORANGE_LIGHTER',
+    'ORANGE',
+    'ORANGE_DARKER',
+    'ORANGE_DARKEST',
+    'RED_LIGHTEST',
+    'RED_LIGHTER',
+    'RED',
+    'RED_DARKER',
+    'RED_DARKEST',
+    'MAGENTA_LIGHTEST',
+    'MAGENTA_LIGHTER',
+    'MAGENTA',
+    'MAGENTA_DARKER',
+    'MAGENTA_DARKEST',
+  ]).optional(),
+  name: z.string().optional(),
+});
+
+export type JiraLabelProperties = z.infer<typeof JiraLabelPropertiesSchema>;
