+++ src/cloud/models/jiraLabelsField.ts
@@ -1,12 +1,12 @@
 import { z } from 'zod';
 import { apiObject, openEnum } from '#/core';
-import { JiraLabelPropertiesInputJackson1Schema } from './jiraLabelPropertiesInputJackson1';
+import { JiraLabelPropertiesSchema } from './jiraLabelProperties';
 import { JiraLabelsInputSchema } from './jiraLabelsInput';
 
 export const JiraLabelsFieldSchema = apiObject({
   bulkEditMultiSelectFieldOption: openEnum(['ADD', 'REMOVE', 'REPLACE', 'REMOVE_ALL']),
   fieldId: z.string(),
-  labelProperties: z.array(JiraLabelPropertiesInputJackson1Schema).optional(),
+  labelProperties: z.array(JiraLabelPropertiesSchema).optional(),
   labels: z.array(JiraLabelsInputSchema),
 });
