+++ src/cloud/models/linkIssueRequestJson.ts
@@ -1,14 +1,13 @@
-import type { z } from 'zod';
-import { apiObject } from '#/core';
-import { CommentSchema } from './comment';
-import { LinkedIssueSchema } from './linkedIssue';
-import { IssueLinkTypeSchema } from './issueLinkType';
+import { LinkIssueRequestSchema, type LinkIssueRequest } from './linkIssueRequest';
 
-export const LinkIssueRequestJsonSchema = apiObject({
-  comment: CommentSchema.optional(),
-  inwardIssue: LinkedIssueSchema,
-  outwardIssue: LinkedIssueSchema,
-  type: IssueLinkTypeSchema,
-});
+/**
+ * @deprecated Renamed to `LinkIssueRequestSchema`, which describes the same shape. This alias is removed in the next
+ *   major version.
+ */
+export const LinkIssueRequestJsonSchema = LinkIssueRequestSchema;
 
-export type LinkIssueRequestJson = z.infer<typeof LinkIssueRequestJsonSchema>;
+/**
+ * @deprecated Renamed to `LinkIssueRequest`, which describes the same shape. This alias is removed in the next major
+ *   version.
+ */
+export type LinkIssueRequestJson = LinkIssueRequest;
