+++ src/cloud/models/page2GetFieldAssociationSchemeResponse.ts
@@ -1,13 +1,16 @@
-import { pageSchema, type Page } from './page';
 import {
-  GetFieldAssociationSchemeResponseSchema,
-  type GetFieldAssociationSchemeResponse,
-} from './getFieldAssociationSchemeResponse';
-
-export const Page2GetFieldAssociationSchemeResponseSchema = pageSchema(GetFieldAssociationSchemeResponseSchema);
+  PageGetFieldAssociationSchemeResponseSchema,
+  type PageGetFieldAssociationSchemeResponse,
+} from './pageGetFieldAssociationSchemeResponse';
 
 /**
- * @deprecated Use `Page<GetFieldAssociationSchemeResponse>`, which describes the same shape. This alias is removed in
- *   the next major version.
+ * @deprecated Renamed to `PageGetFieldAssociationSchemeResponseSchema`, which describes the same shape. This alias is
+ *   removed in the next major version.
  */
-export type Page2GetFieldAssociationSchemeResponse = Page<GetFieldAssociationSchemeResponse>;
+export const Page2GetFieldAssociationSchemeResponseSchema = PageGetFieldAssociationSchemeResponseSchema;
+
+/**
+ * @deprecated Renamed to `PageGetFieldAssociationSchemeResponse`, which describes the same shape. This alias is removed
+ *   in the next major version.
+ */
+export type Page2GetFieldAssociationSchemeResponse = PageGetFieldAssociationSchemeResponse;
