+++ src/cloud/models/page2GetProjectsWithFieldSchemesResponse.ts
@@ -1,13 +1,16 @@
-import { pageSchema, type Page } from './page';
 import {
-  GetProjectsWithFieldSchemesResponseSchema,
-  type GetProjectsWithFieldSchemesResponse,
-} from './getProjectsWithFieldSchemesResponse';
-
-export const Page2GetProjectsWithFieldSchemesResponseSchema = pageSchema(GetProjectsWithFieldSchemesResponseSchema);
+  PageGetProjectsWithFieldSchemesResponseSchema,
+  type PageGetProjectsWithFieldSchemesResponse,
+} from './pageGetProjectsWithFieldSchemesResponse';
 
 /**
- * @deprecated Use `Page<GetProjectsWithFieldSchemesResponse>`, which describes the same shape. This alias is removed in
- *   the next major version.
+ * @deprecated Renamed to `PageGetProjectsWithFieldSchemesResponseSchema`, which describes the same shape. This alias is
+ *   removed in the next major version.
  */
-export type Page2GetProjectsWithFieldSchemesResponse = Page<GetProjectsWithFieldSchemesResponse>;
+export const Page2GetProjectsWithFieldSchemesResponseSchema = PageGetProjectsWithFieldSchemesResponseSchema;
+
+/**
+ * @deprecated Renamed to `PageGetProjectsWithFieldSchemesResponse`, which describes the same shape. This alias is
+ *   removed in the next major version.
+ */
+export type Page2GetProjectsWithFieldSchemesResponse = PageGetProjectsWithFieldSchemesResponse;
