+++ src/cloud/models/pageGetFieldAssociationSchemeResponse.ts
@@ -1,0 +1,13 @@
+import { pageSchema, type Page } from './page';
+import {
+  GetFieldAssociationSchemeResponseSchema,
+  type GetFieldAssociationSchemeResponse,
+} from './getFieldAssociationSchemeResponse';
+
+export const PageGetFieldAssociationSchemeResponseSchema = pageSchema(GetFieldAssociationSchemeResponseSchema);
+
+/**
+ * @deprecated Use `Page<GetFieldAssociationSchemeResponse>`, which describes the same shape. This alias is removed in
+ *   the next major version.
+ */
+export type PageGetFieldAssociationSchemeResponse = Page<GetFieldAssociationSchemeResponse>;
